fix(ci): line-limit maintainability — split two oversized unit test modules over 1500 lines
